fix(review): archi fichiers landing + server components + fix ui divers
