--- ui/app/(public)/(generique)/_components/footer-cta-section.module.scss
@@ -1,174 +1,0 @@
-@use "@/styles/breakpoints" as *;
-@use "../../landing-shared" as *;
-
-.section {
-  width: 100%;
-}
-
-.illustrationBand {
-  position: relative;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  width: 100%;
-  height: 13rem;
-  overflow: visible;
-  background: linear-gradient(0deg, #adadf9 50%, rgba(173, 173, 249, 0) 80%);
-
-  @include respond-from(md) {
-    height: 13rem;
-  }
-}
-
-.illustrationContainer {
-  max-width: var(--container-xl);
-}
-
-.illustration {
-  position: relative;
-  z-index: 1;
-  display: block;
-  width: 100%;
-  height: auto;
-  margin-bottom: -1.375rem;
-  user-select: none;
-
-  @include respond-from(md) {
-    margin-bottom: -2.5rem;
-  }
-}
-
-.ctaBand {
-  width: 100%;
-  padding-top: 2.5rem;
-  padding-right: 2rem;
-  padding-bottom: 2.5rem;
-  padding-left: 2rem;
-  text-align: left;
-  background: var(--light-decisions-artwork-artwork-major-blue-france, #000091);
-
-  @include respond-from(md) {
-    padding-top: 4rem;
-    padding-right: 2.5rem;
-    padding-bottom: 4rem;
-    padding-left: 2.5rem;
-  }
-
-  @include respond-from(lg) {
-    padding-top: 5rem;
-    padding-right: 7.5rem;
-    padding-bottom: 5rem;
-    padding-left: 7.5rem;
-  }
-}
-
-.srOnly {
-  @extend %landing-sr-only;
-}
-
-.ctaContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 3.125rem;
-  justify-content: space-between;
-  max-width: 996px;
-  margin-top: 0;
-  margin-right: auto;
-  margin-bottom: 0;
-  margin-left: auto;
-
-  @include respond-from(md) {
-    flex-direction: row;
-    gap: 2rem;
-  }
-
-  @include respond-from(lg) {
-    gap: 2rem;
-  }
-}
-
-.ctaItem {
-  display: flex;
-  flex-direction: column;
-  gap: 1.25rem;
-  align-items: center;
-  width: 100%;
-  max-width: 320px;
-  margin-top: 0;
-  margin-right: auto;
-  margin-bottom: 0;
-  margin-left: auto;
-
-  @include respond-from(md) {
-    flex: 1;
-    align-items: flex-start;
-    width: auto;
-    max-width: none;
-    margin: 0;
-  }
-
-  @include respond-from(lg) {
-    flex: 0 0 auto;
-    width: 280px;
-  }
-}
-
-.ctaTop {
-  display: flex;
-  flex-direction: column;
-  gap: 0.875rem;
-  align-items: center;
-  width: 100%;
-
-  @include respond-from(md) {
-    align-items: flex-start;
-  }
-}
-
-.ctaIcon {
-  font-size: 32px;
-  color: var(--light-decisions-background-background-default-grey, #fff);
-}
-
-.ctaTitle {
-  margin: 0;
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 28px;
-  color: var(--light-decisions-background-background-default-grey, #fff);
-  text-align: center;
-
-  @include respond-from(md) {
-    text-align: left;
-  }
-}
-
-.ctaActions {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  height: 100%;
-
-  @include respond-from(md) {
-    align-items: flex-end;
-  }
-}
-
-.ctaButton {
-  justify-content: center;
-  width: 100%;
-  margin-top: auto;
-  margin-right: auto;
-  margin-bottom: 0;
-  margin-left: auto;
-  color: var(--light-decisions-artwork-artwork-major-blue-france, #000091);
-  background-color: var(--light-decisions-background-background-default-grey, #fff);
-
-  &:hover {
-    background-color: var(--light-decisions-background-background-alt-grey-hover, #dfdfdf);
-  }
-}
-
-.ctaLink {
-  color: var(--light-decisions-background-background-default-grey, #fff) !important;
-}
