+++ ui/app/(public)/_components/FooterCTASection.tsx
@@ -5,52 +5,82 @@ import { DsfrLink } from "@/app/_components/link/DsfrLink";
 
 import styles from "./footer-cta-section.module.scss";
 
-type LinkInscription = "/missions_locales" | "/operateur_public" | "/organisme_formation";
+type CTA = {
+  iconClass: string;
+  title: string;
+  buttonLabel: string;
+  href: string;
+};
 
-export function FooterCTASection({ linkInscription = "" }: { linkInscription?: LinkInscription }) {
+const CTAS: Array<CTA> = [
+  {
+    iconClass: "ri-school-line",
+    title: "Je suis un établissement de formation (CFA)",
+    buttonLabel: "Créer mon compte",
+    href: "/cfa",
+  },
+  {
+    iconClass: "fr-icon-community-line",
+    title: "Je suis une Mission Locale",
+    buttonLabel: "Créer mon compte",
+    href: "/missions-locales",
+  },
+  {
+    iconClass: "fr-icon-government-line",
+    title: "Je suis un-e référent-e territorial•e",
+    buttonLabel: "Obtenir mon accès",
+    href: "/territoires",
+  },
+];
+
+function CTAItem(cta: CTA) {
+  return (
+    <div className={styles.ctaItem}>
+      <div className={styles.ctaTop}>
+        <span className={`${cta.iconClass} ${styles.ctaIcon}`} aria-hidden="true" />
+        <h3 className={styles.ctaTitle}>{cta.title}</h3>
+      </div>
+      <div className={styles.ctaActions}>
+        <Button
+          className={styles.ctaButton}
+          iconId="fr-icon-arrow-right-line"
+          iconPosition="right"
+          linkProps={{ href: cta.href, "aria-label": `${cta.buttonLabel} : ${cta.title}` }}
+        >
+          {cta.buttonLabel}
+        </Button>
+        <DsfrLink
+          className={styles.ctaLink}
+          href={cta.href}
+          aria-label={`Découvrir les fonctionnalités : ${cta.title}`}
+        >
+          Découvrir les fonctionnalités
+        </DsfrLink>
+      </div>
+    </div>
+  );
+}
+
+export function FooterCTASection() {
   return (
     <section className={styles.section}>
       <div className={styles.illustrationBand}>
         <div className={styles.illustrationContainer}>
           <Image
             className={styles.illustration}
-            src="/images/home/illustration-footer-cfa-ml.png"
+            src="/images/home/illustration-footer.png"
             alt=""
-            width={494}
-            height={185}
+            width={1264}
+            height={384}
           />
         </div>
       </div>
       <div className={styles.ctaBand}>
+        <h2 className={styles.srOnly}>Rejoindre le service</h2>
         <div className={styles.ctaContainer}>
-          <div>
-            <p className={styles.subtitle}>
-              Les Missions Locales de votre territoire sont prêtes à accompagner vos apprenants en difficulté.
-            </p>
-            <p className={styles.title}>Commencez à collaborer sur le Tableau de bord de l’apprentissage.</p>
-          </div>
-          <div className={styles.actions}>
-            <Button
-              className={styles.primaryButton}
-              priority="primary"
-              iconId="fr-icon-arrow-right-line"
-              iconPosition="right"
-              linkProps={{
-                href: `/auth/inscription${linkInscription}`,
-                "aria-label":
-                  "Inscrire mon établissement : Commencez à collaborer sur le Tableau de bord de l’apprentissage.",
-              }}
-            >
-              Inscrire mon établissement
-            </Button>
-            <DsfrLink
-              href={`/auth/inscription${linkInscription}`}
-              className={styles.demoLink}
-              aria-label="Demander une démo : Commencez à collaborer sur le Tableau de bord de l’apprentissage."
-            >
-              Demander une démo
-            </DsfrLink>
-          </div>
+          {CTAS.map((cta) => (
+            <CTAItem key={cta.href} {...cta} />
+          ))}
         </div>
       </div>
     </section>
