+++ ui/app/(public)/_components/SponsorsSection.tsx
@@ -1,41 +1,23 @@
-import Image from "next/image";
+import { BaseSponsorsSection, type Sponsor } from "./shared/BaseSponsorsSection";
 
-export type Sponsor = {
-  image: { src: string; width: number; height: number } | null;
-  title: string;
-  description: string;
-};
+const SPONSORS: Array<Sponsor> = [
+  {
+    image: { src: "/images/home/ml-actives.png", width: 360, height: 246 },
+    title: "+ 200 Missions Locales actives",
+    description: "Un réseau de Missions Locales déjà actives sur le service",
+  },
+  {
+    image: { src: "/images/home/made-in-DGEFP.png", width: 360, height: 200 },
+    title: "Un service proposé par la DGEFP",
+    description: "Un service co-construit avec les acteurs terrain CFA et Missions Locales directement",
+  },
+  {
+    image: null,
+    title: "+ XXX établissements de formation",
+    description: "Les CFA choisissent le Tableau de bord pour collaborer avec les Missions Locales.",
+  },
+];
 
-function SponsorItem({ image, title, description, styles }: Sponsor & { styles: Record<string, string> }) {
-  return (
-    <div className={styles.item}>
-      {image ? (
-        <Image
-          className={styles.image}
-          src={image.src}
-          alt=""
-          width={image.width}
-          height={image.height}
-          sizes="(max-width: 768px) 100px, 100px"
-        />
-      ) : (
-        <div className={styles.image} aria-hidden="true" />
-      )}
-      <div className={styles.textContainer}>
-        <p className={styles.title}>{title}</p>
-        <p className={styles.description}>{description}</p>
-      </div>
-    </div>
-  );
-}
-
-export function SponsorsSection({ sponsors, styles }: { sponsors: Sponsor[]; styles: Record<string, string> }) {
-  return (
-    <section className={styles.section} aria-label="Chiffres clés du service">
-      <h2 className={styles.srOnly}>Le service en chiffres</h2>
-      {sponsors.map((sponsor) => (
-        <SponsorItem key={sponsor.title} {...sponsor} styles={styles} />
-      ))}
-    </section>
-  );
+export function SponsorsSection() {
+  return <BaseSponsorsSection sponsors={SPONSORS} />;
 }
