+++ ui/app/(public)/_components/shared/footer-cta-section.module.scss
@@ -1,0 +1,145 @@
+@use "@/styles/breakpoints" as *;
+@use "../../landing-shared" as *;
+
+.section {
+  @extend %landing-section;
+  gap: 0;
+  width: 100%;
+  padding: 0;
+}
+
+.illustrationBand {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  overflow: visible;
+  background: linear-gradient(0deg, #adadf9 -81.57%, rgba(173, 173, 249, 0) 96.71%);
+}
+
+.illustrationContainer {
+  position: relative;
+  display: flex;
+  align-items: flex-end;
+  justify-content: flex-end;
+  width: 100%;
+  max-width: var(--container-xl);
+}
+
+.illustration {
+  position: relative;
+  z-index: 1;
+  display: block;
+  max-inline-size: 100%;
+  height: auto;
+  padding-right: 16px;
+  padding-left: 16px;
+  margin-bottom: -36px;
+  user-select: none;
+
+  @include respond-from(md) {
+    margin-bottom: -40px;
+  }
+}
+
+.ctaBand {
+  width: 100%;
+  padding-top: 40px;
+  padding-right: 24px;
+  padding-bottom: 40px;
+  padding-left: 24px;
+  text-align: left;
+  background: var(--light-decisions-artwork-artwork-major-blue-france, #000091);
+
+  @include respond-from(md) {
+    padding-top: 64px;
+    padding-right: 40px;
+    padding-bottom: 64px;
+    padding-left: 40px;
+  }
+
+  @include respond-from(lg) {
+    padding-top: 64px;
+    padding-right: 120px;
+    padding-bottom: 64px;
+    padding-left: 120px;
+  }
+}
+
+.ctaContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  justify-content: center;
+  max-width: var(--container-xl);
+  margin-top: 0;
+  margin-right: auto;
+  margin-bottom: 0;
+  margin-left: auto;
+
+  @include respond-from(md) {
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+  }
+}
+
+.subtitle {
+  max-width: 588px;
+  font-size: 20px;
+  font-weight: 700;
+  line-height: 28px;
+  color: var(--light-decisions-background-background-default-grey, #fff);
+
+  @include respond-from(md) {
+    margin-bottom: 32px;
+    font-size: 20px;
+    line-height: 32px;
+  }
+}
+
+.title {
+  max-width: 588px;
+  margin: 0;
+  font-size: 32px;
+  font-weight: 700;
+  line-height: 40px;
+  color: var(--light-decisions-background-background-default-grey, #fff);
+
+  @include respond-from(md) {
+    font-size: 32px;
+    line-height: 40px;
+  }
+}
+
+.actions {
+  display: flex;
+  flex-shrink: 0;
+  flex-direction: column;
+  gap: 12px;
+
+  @include respond-from(md) {
+    flex-direction: column;
+    gap: 16px;
+    align-items: flex-end;
+    justify-content: end;
+    text-align: right;
+  }
+}
+
+.primaryButton {
+  color: var(--light-decisions-artwork-artwork-major-blue-france, #000091);
+  background-color: var(--light-decisions-background-background-default-grey, #fff);
+
+  &:hover {
+    background-color: var(--light-decisions-background-background-alt-grey-hover, #dfdfdf);
+  }
+}
+
+.demoLink {
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+  color: var(--light-decisions-background-background-default-grey, #fff) !important;
+  text-decoration: underline !important;
+}
