+++ ui/app/(public)/_components/shared/pause-button.module.scss
@@ -1,5 +1,5 @@
 @use "@/styles/breakpoints" as *;
-@use "../landing-shared" as *;
+@use "../../landing-shared" as *;
 
 .pauseButton {
   @extend %landing-action-button;
