+++ ui/app/(public)/_components/shared/temoignages-section.module.scss
@@ -1,5 +1,5 @@
 @use "@/styles/breakpoints" as *;
-@use "../landing-shared" as *;
+@use "../../landing-shared" as *;
 
 .section {
   @extend %landing-section;
