--- ui/app/(public)/accueil-mission-locale/_components/processus-section.module.scss
@@ -1,182 +1,0 @@
-@use "@/styles/breakpoints" as *;
-@use "../../landing-shared" as *;
-
-.section {
-  @extend %landing-section;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  text-align: left;
-  background: var(--light-decisions-background-background-alt-grey, #f6f6f6);
-}
-
-.contentWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  width: 100%;
-  max-width: var(--container-xl);
-}
-
-.intro {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  @include respond-from(lg) {
-    flex-direction: row;
-    align-items: flex-start;
-    justify-content: space-between;
-  }
-}
-
-.title {
-  flex: 1;
-  margin: 0;
-  font-size: 24px;
-  font-weight: 700;
-  line-height: 2rem;
-  color: var(--light-decisions-text-text-title-grey, #161616);
-
-  @include respond-from(md) {
-    max-width: 690px;
-    font-size: 2rem;
-    line-height: 40px;
-  }
-}
-
-.subtitle {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  color: var(--light-decisions-text-text-mention-grey, #666);
-
-  @include respond-from(lg) {
-    flex-shrink: 0;
-    max-width: 382px;
-  }
-}
-
-.srOnly {
-  @extend %landing-sr-only;
-}
-
-.carousel {
-  width: 100%;
-  overflow: visible;
-  --card-width: calc(100vw - 2rem);
-  --card-gap: 16px;
-
-  @include respond-from(md) {
-    --card-width: 240px;
-    --card-gap: 24px;
-  }
-
-  @include respond-from(lg) {
-    --card-width: 282px;
-    --card-gap: 24px;
-  }
-}
-
-.track {
-  display: flex;
-  gap: var(--card-gap);
-  transform: translateX(
-    calc(50vw - (var(--card-width) / 2) - var(--position, 0) * (var(--card-width) + var(--card-gap)))
-  );
-  will-change: transform;
-}
-
-.card {
-  display: flex;
-  flex: 0 0 auto;
-  flex-direction: column;
-  gap: 0;
-  align-items: center;
-  width: var(--card-width);
-  min-height: 300px;
-  padding: 2rem;
-  text-align: left;
-  background: var(--light-decisions-background-background-default-grey, #fff);
-  border: 1px solid var(--light-decisions-border-default-grey, #ddd);
-  border-bottom: 4px solid var(--light-decisions-border-contrast-grey, #929292);
-}
-
-.cardImageWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 120px;
-  height: 80px;
-  margin-bottom: 16px;
-  overflow: hidden;
-}
-
-.cardImage {
-  width: 120px;
-  height: 80px;
-  object-fit: contain;
-}
-
-.cardTitle {
-  width: 100%;
-  margin: 0;
-  font-size: 20px;
-  font-weight: 400;
-  line-height: 28px;
-  color: var(--light-decisions-text-text-title-grey, #161616);
-  text-align: left;
-}
-
-.progress {
-  display: flex;
-  gap: 8px;
-}
-
-.progressStep,
-.progressStepActive {
-  flex: 1 0 0;
-  height: 8px;
-  transition: background-color 0.3s ease;
-}
-
-.progressStep {
-  background: var(--light-decisions-background-background-contrast-grey, #eee);
-}
-
-.progressStepActive {
-  background: var(--light-decisions-background-background-action-high-blue-france, #000091);
-}
-
-.controls {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.iconButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 8px;
-  color: var(--light-decisions-text-text-action-high-blue-france, #000091);
-  cursor: pointer;
-  background: none;
-  border: none;
-
-  &:hover {
-    background: rgba(0, 0, 145, 0.08);
-  }
-}
-
-.controlsRight {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-}
-
-.cta {
-  display: flex;
-  justify-content: center;
-}
